Move xtrigger_mgr fixture to conftest.py and fix test_config.py
